Add installed usage and sf2e weapon/armor upgrades
