+++ src/module/item/base/data/values.ts
@@ -20,7 +20,7 @@ class MystifiedTraits {
     }
 }
 
-const ITEM_CARRY_TYPES = ["attached", "dropped", "held", "stowed", "worn", "implanted"] as const;
+const ITEM_CARRY_TYPES = ["installed", "attached", "dropped", "held", "stowed", "worn", "implanted"] as const;
 
 const RANGE_TRAITS = [
     "range-5",
