+++ src/module/item/physical/sheet.ts
@@ -3,11 +3,10 @@ import type { FormSelectOption } from "@client/applications/forms/fields.d.mts";
 import type { AppV1RenderOptions } from "@client/appv1/api/application-v1.d.mts";
 import type { PhysicalItemPF2e } from "@item";
 import { ItemSheetDataPF2e, ItemSheetOptions, ItemSheetPF2e } from "@item/base/sheet/sheet.ts";
-import { getAdjustment } from "@module/sheet/helpers.ts";
+import { getAdjustment, isControlDown } from "@module/sheet/helpers.ts";
 import { TextEditorPF2e } from "@system/text-editor.ts";
 import { ErrorPF2e, htmlClosest, htmlQuery, localizer, sortStringRecord, tupleHasValue } from "@util";
 import * as R from "remeda";
-import { detachSubitem } from "./helpers.ts";
 import { CoinsPF2e, MaterialValuationData } from "./index.ts";
 import { PRECIOUS_MATERIAL_GRADES } from "./values.ts";
 
@@ -172,7 +171,7 @@ class PhysicalItemSheetPF2e<TItem extends PhysicalItemPF2e> extends ItemSheetPF2
                 case "edit-subitem":
                     return subitem.sheet.render(true);
                 case "detach-subitem":
-                    return detachSubitem(subitem, event.ctrlKey);
+                    return subitem.detach({ skipConfirm: isControlDown(event) });
                 case "delete-subitem": {
                     return event.ctrlKey ? subitem.delete() : subitem.deleteDialog();
                 }
