+++ static/lang/en.json
@@ -2373,6 +2373,11 @@
                 "ShowTokenIcon": "Show token icon?",
                 "Unidentified": "Unidentified?"
             },
+            "Equipment": {
+                "Category": {
+                    "upgrade": "Upgrade"
+                }
+            },
             "FIELDS": {
                 "category": {
                     "label": "Category"
@@ -4804,6 +4809,17 @@
         "TraitIkon": "Ikon",
         "TraitIllusion": "Illusion",
         "TraitImplanted": "Implanted",
+        "TraitInstalledInAGrenadeLauncherOrTwoHandedWeaponWithAnUndermoundedGrenadeLauncher": "Installed in a grenade launcher or two-handed weapon with an undermounted grenade launcher",
+        "TraitInstalledInArmor": "Installed in armor",
+        "TraitInstalledInArmorWithTheEnergyShieldingUpgrade": "Installed in armor with the energy shielding upgrade",
+        "TraitInstalledInArmorWithTheExposedTrait": "Installed in armor with the exposed trait",
+        "TraitInstalledInAWeapon": "Installed in a weapon",
+        "TraitInstalledInAWeaponSight": "Installed in a weapon (sight)",
+        "TraitInstalledInAWeaponWithTheKickbackTrait": "Installed in a weapon with the kickback trait",
+        "TraitInstalledInOneHandedWeaponGrip": "Installed in one-handed weapon (grip)",
+        "TraitInstalledInRangedWeaponWithoutALoudener": "Installed in ranged weapon without a loudener",
+        "TraitInstalledInTwoHandedWeapon": "Installed in two-handed weapon",
+        "TraitInstalledOnAWeaponWithoutASilencer": "Installed on a weapon without a silencer",
         "TraitImpulse": "Impulse",
         "TraitIncapacitation": "Incapacitation",
         "TraitIncarnate": "Incarnate",
