+++ static/templates/items/item-attacher.hbs
@@ -16,7 +16,7 @@
 
 <footer>
     <button type="button" data-action="pick" value="0">{{localize "PF2E.Item.Physical.Attach.Label"}}</button>
-    {{#if item.actor.skills.crafting}}
+    {{#if requiresCrafting}}
         <label>
             <input type="checkbox" data-crafting-check {{checked (not user.isGM)}} />
             <span>{{localize "PF2E.Item.Physical.Attach.CraftingCheck"}}</span>
